Split IndexMapCore to its own module
