+++ src/lib.rs
@@ -79,6 +79,9 @@ mod mutable_keys;
 mod serde;
 mod util;
 
+#[macro_use]
+mod map_core;
+
 pub mod map;
 pub mod set;
